Added a python verions of sensor generator

+++ Code/raspberry/kill_spawned.py
@@ -6,6 +6,8 @@ SPAWNED_PROCESSES_FNAME = "spawned-processes"
 if not os.path.isfile(SPAWNED_PROCESSES_FNAME):
     raise OSError("A file, '" + SPAWNED_PROCESSES_FNAME + "', does not exist.")
 
+bdelete = True
+
 with open(SPAWNED_PROCESSES_FNAME, 'r') as fp:
     line = fp.readline()
     while line:
@@ -14,9 +16,12 @@ with open(SPAWNED_PROCESSES_FNAME, 'r') as fp:
         except OSError as err:
             if err.errno == 1:
                 print("Please run script as admin (sudo)")
-                raise
-            print("Killing process error: " + line + " - " + str(err))
+                bdelete = False
+                break
+        except Exception as err:
+            print("Error killing process: " + line + " - " + str(err))
             pass
         line = fp.readline()
 
-os.remove(SPAWNED_PROCESSES_FNAME)
+if bdelete:
+    os.remove(SPAWNED_PROCESSES_FNAME)
